+++ pack.py
@@ -1,0 +1,28 @@
+# This Python File Handles the Pack Object of ModDude!
+# Path: pack.py
+
+# Import Modules
+
+class Pack:
+    url = ""
+    pack_name = ""
+    pack_version = ""
+    game = ""
+    game_version = ""
+    mod_loader = ""
+    mod_loader_version = ""
+    pack_urls = []
+    mods = []
+    banner_url = ""
+
+    def __init__(self):
+        self.url = ""
+        self.pack_name = ""
+        self.pack_version = ""
+        self.game = ""
+        self.game_version = ""
+        self.mod_loader = ""
+        self.mod_loader_version = ""
+        self.pack_urls = []
+        self.mods = []
+        self.banner_url = ""
